# Conflicts:
#	src/main/java/com/pirimidtech/ptp/entity/CompanyDetail.java
#	src/main/java/com/pirimidtech/ptp/entity/MutualFundCategory.java
#	src/main/java/com/pirimidtech/ptp/entity/MutualFundDetail.java
#	src/main/java/com/pirimidtech/ptp/entity/MutualFundOrder.java
#	src/main/java/com/pirimidtech/ptp/entity/MutualFundStatistic.java
#	src/main/java/com/pirimidtech/ptp/entity/Position.java
#	src/main/java/com/pirimidtech/ptp/entity/StockBrands.java
#	src/main/java/com/pirimidtech/ptp/entity/StockDetail.java
#	src/main/java/com/pirimidtech/ptp/entity/StockStatistic.java
#	src/main/java/com/pirimidtech/ptp/entity/User.java
#	src/main/resources/application.properties

+++ src/main/java/com/pirimidtech/ptp/entity/StockPrice.java
@@ -5,7 +5,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.persistence.JoinColumn;
+
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -23,7 +28,7 @@ public class StockPrice {
     @Id
     private UUID stockPriceId;
 
-    private Integer  price;
+    private float  price;
 
     private LocalDateTime timestamp;
